refactor dimensions
