+++ src/lib/Dimensions.svelte
@@ -1,7 +1,21 @@
-<script lang="ts">
-	import type {Dimensions} from '$lib/dimensions';
+<script lang="ts" context="module">
+	import {getContext, setContext} from 'svelte';
 	import type {Writable} from 'svelte/store';
 
+	export interface Dimensions {
+		width: number;
+		height: number;
+	}
+
+	const KEY = Symbol();
+
+	export const getDimensions = (): Writable<Dimensions> => getContext(KEY);
+
+	export const setDimensions = (dimensions: Writable<Dimensions>): Writable<Dimensions> =>
+		setContext(KEY, dimensions);
+</script>
+
+<script lang="ts">
 	export let dimensions: Writable<Dimensions>;
 
 	// This measures an onscreen element instead of window.innerWidth/Height
